debug key event

+++ src/scrcpy.rs
@@ -743,11 +743,11 @@ async fn tsk_scrcpy_control(
                                 let key_down = key_ev.down();
                                 key_code=key_ev.keycode();
                                 //remap fast forward/rewind
-                                if key_code == KeyCode::KEYCODE_MEDIA_FAST_FORWARD {
-                                    key_code=KeyCode::KEYCODE_L;
+                                if key_code == KeyCode::KEYCODE_MEDIA_FAST_FORWARD as u32 {
+                                    key_code=KeyCode::KEYCODE_L as i32;
                                 }
-                                else if key_code == KeyCode::KEYCODE_MEDIA_REWIND {
-                                    key_code=KeyCode::KEYCODE_J;
+                                else if key_code == KeyCode::KEYCODE_MEDIA_REWIND as u32 {
+                                    key_code=KeyCode::KEYCODE_J as i32;
                                 }
                                 let mut _action: u8;
                                 if key_down
